Implemented colours via pgfkeys
Finally sat down and learnt tikz. Maked to get the rectangles and logo
positioned nice and clean like.

Only for 16:9 at the moment, but it’s just some if/elses and some maths
to make it valid for all ratios.

Struggled, but learnt pgfkeys colours can now be implemented for the
theme via the `themecolor=<color>` option (themecolour also permissible)

TODO:
 - title page
 - mono option
 - general beamer options (fg, bg, etc)