Custom taglines are now possible
Added custom tailgates. Tagline can be set with \tagline{} default is
still PERSONALISED EXPERIENCES: WORLD-CLASS RESULTS.

Changed box of section slides to be left aligned on page rather tan
centred and increased the length of the progress bar to the width of
the brand mark box.

Added \tagline{} to the documentation.